relatedEntries.js: dockerApiUrl fix
Downloaded the latest version of the code and was having trouble running the app, specifically dockerApiUrl returned the correct value initially, but then returned 'null' when called in the sendApiKey function. I think the reason for this was the timing of the execution of the code, since the sendApiKey function is asynchronous.

This fix ensures that the dockerApiBaseUrl is set correctly before being accessed, which avoids the 'TypeError' that results from the 'null' value of document.currentScript.

I also had to update all instances of 'dockerApiUrl' to 'dockerApiUrl()'.

This code has been tested and works properly now.