fullTitle: Language agnostic code
This change should make the code for getting the 'fullTitle' object in OJS JSON language-agnostic. It extracts the first publication's data, then tries to determine the appropriate language to use for the title by checking the "locale" variable at the submission level. If the specified language is available in the "fullTitle" object, it selects the corresponding title; otherwise, it falls back to the default language "en_US" or any available language title.

I have tested this code and it doesn't break anything (although I feel like maybe it makes it slightly slower to load, though hard to say), but I haven't tried to switch the submission language, since I'm not sure how to do that in OJS.